gemini_program/requirements.txt
Flask
google-generativeai
python-dotenv
authlib
requests
pdfplumber
python-docx
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
flask
PyPDF2
